Project_Mantra
Objective of the Project:-
---------------------------
The idea aims at the Safety of women from the increasing violence raising against them in the society. At this point of time, Women are required to use utmost of the Technology to tackle the assault on them. With that intent, through the Project_Mantra i have designed and developed a Safety device to overcome the assault on them. As the technology is growing up, the way of putting down the violence should also be in a progressive manner, so that Women who are in a tragic situation should be able to escape from the Violence area or make the criminal run away in order to safeguard their lives.

Mantra - Process Flow:-
--------------------------
Women of every locality gets registered with the Ministry of Women welfare of the Nation, for obtaining the IoT device, Mantra. Finger impression of the owner is registered against every device. 

This medical IoT gadget would be designed as a neck accessory which any woman can wear on her own with ease of carrying it everywhere.

The IoT gadget with the Touch ID sensor gets activated with the Finger print identification of the owner(s) of the gadget. Activation of the gadget sends an alert to the local police station along 
with the location and timestamp, recording it as an incident.

Pre-loaded drug patches arranged in FIFO order serves as a bullet module. 

Drugs can be either of the below, based on the approval of the Healthcare ministry of the nation:
 (i) Anesthesia - To induce drowsiness to the criminal to refrain from proceeding violence against women.
 (ii) Endothelin-1(ET-1) - To cause itching in human cell and furthermore evokes scratching behavior.

A handy needless injector device paired with Google Project Soli will involve in injecting the drug which along with the nano GPS module to the Epidermis of the criminal's human body. Nano GPS serves the purpose of tracking the navigation path of the criminal wherever he traverse.